system passes continues

+++ citrus/dynamics/geom.h
@@ -88,7 +88,7 @@ namespace citrus::geom {
 			a = glm::dot(edge1, h);
 			if(a > -EPSILON && a < EPSILON)
 				return hit(false, 0.0f);    // This ray is parallel to this triangle.
-			f = 1.0 / a;
+			f = float(1.0 / a);
 			s = ray.p - vertex0;
 			u = f * glm::dot(s, h);
 			if(u < 0.0 || u > 1.0)
