+++ citrus/graphics/window.cpp
@@ -234,7 +234,7 @@ namespace citrus::graphics {
 	void window::present(uint32_t index, vector<VkSemaphore> waits) {
 		VkPresentInfoKHR presentInfo = { };
 		presentInfo.sType = VK_STRUCTURE_TYPE_PRESENT_INFO_KHR;
-		presentInfo.waitSemaphoreCount = waits.size();
+		presentInfo.waitSemaphoreCount = (uint32_t) waits.size();
 		presentInfo.pWaitSemaphores = waits.data();
 		presentInfo.swapchainCount = 1;
 		presentInfo.pSwapchains = &_inst->_swapChain;
